+++ packages/alfa-act/src/audit.ts
@@ -8,26 +8,23 @@ import { Oracle } from "./oracle";
 import { Outcome } from "./outcome";
 import { Rule } from "./rule";
 
-export class Audit<I, T = unknown, Q = unknown> {
-  public static of<I, T = unknown, Q = unknown>(
+export class Audit<I, T = unknown, Q = never> {
+  public static of<I, T = unknown, Q = never>(
     input: I,
+    rules: Iterable<Rule<I, T, Q>>,
     oracle: Oracle<Q> = () => Future.now(None)
   ): Audit<I, T, Q> {
-    return new Audit(input, oracle, List.empty());
+    return new Audit(input, List.from(rules), oracle);
   }
 
   private readonly _input: I;
-  private readonly _oracle: Oracle<Q>;
   private readonly _rules: List<Rule<I, T, Q>>;
+  private readonly _oracle: Oracle<Q>;
 
-  private constructor(input: I, oracle: Oracle<Q>, rules: List<Rule<I, T, Q>>) {
+  private constructor(input: I, rules: List<Rule<I, T, Q>>, oracle: Oracle<Q>) {
     this._input = input;
-    this._oracle = oracle;
     this._rules = rules;
-  }
-
-  public add(rule: Rule<I, T, Q>): Audit<I, T, Q> {
-    return new Audit(this._input, this._oracle, this._rules.append(rule));
+    this._oracle = oracle;
   }
 
   public evaluate(): Future<Iterable<Outcome<I, T, Q>>> {