+++ packages/alfa-assert/src/assert.ts
@@ -5,7 +5,7 @@ import { Equatable } from "@siteimprove/alfa-equatable";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { List } from "@siteimprove/alfa-list";
 import { Option } from "@siteimprove/alfa-option";
-import { Rules } from "@siteimprove/alfa-rules";
+import { Rules, Question } from "@siteimprove/alfa-rules";
 import * as web from "@siteimprove/alfa-web";
 
 const { find, reduce } = Iterable;
@@ -50,32 +50,30 @@ export namespace Assert {
 
     export function isAccessible(
       page: web.Page,
-      scope: Iterable<Rule<web.Page, Target, any>> = Rules.values()
+      scope: Iterable<Rule<web.Page, Target, Question>> = Rules.values()
     ): Future<Option<Error<Target>>> {
-      const audit = reduce(
-        scope,
-        (audit, rule) => audit.add(rule),
-        Audit.of(page)
-      );
+      return Audit.of(page, scope)
+        .evaluate()
+        .map((outcomes) =>
+          find(outcomes, (outcome): outcome is Outcome.Failed<
+            web.Page,
+            Target,
+            Question
+          > => Outcome.isFailed(outcome)).map((outcome) => {
+            const { target, expectations } = outcome;
 
-      return audit.evaluate().map((outcomes) =>
-        find(outcomes, (outcome): outcome is Outcome.Failed<web.Page, Target> =>
-          Outcome.isFailed(outcome)
-        ).map((outcome) => {
-          const { target, expectations } = outcome;
+            const reasons = reduce(
+              expectations,
+              (reasons, [id, expectation]) =>
+                expectation.isOk()
+                  ? reasons.append(expectation.get().message)
+                  : reasons,
+              List.empty<string>()
+            );
 
-          const reasons = reduce(
-            expectations,
-            (reasons, [id, expectation]) =>
-              expectation.isOk()
-                ? reasons.append(expectation.get().message)
-                : reasons,
-            List.empty<string>()
-          );
-
-          return Error.of(target, reasons);
-        })
-      );
+            return Error.of(target, reasons);
+          })
+        );
     }
   }
 }