+++ packages/alfa-rules/src/sia-r47/rule.ts
@@ -72,9 +72,10 @@ export namespace Outcomes {
   );
 }
 
-/*
- * Parses a list of "name=value" properties according to
- * https://drafts.csswg.org/css-device-adapt/#parsing-algorithm
+/**
+ * Parses a list of "name=value" properties.
+ *
+ * @see https://drafts.csswg.org/css-device-adapt/#parsing-algorithm
  *
  * @remarks
  * This seems to be the iOS/Safari algorithm and other browsers might handle it
@@ -100,31 +101,40 @@ export function parsePropertiesList(
   let name: string;
   let value: string;
   while (i < length) {
-    // find the start of the next name
+    // Find the start of the next name
     while (i < length && allSpecial.has(propertiesList[i])) {
       i++;
     }
-    // parse the name of the property
+
+    // Parse the name of the property
     start = i;
+
     while (i < length && !allSpecial.has(propertiesList[i])) {
       i++;
     }
+
     name = propertiesList.substring(start, i);
-    // find a separator (end of property) or equal sign
+
+    // Find a separator (end of property) or equal sign
     while (i < length && !separatorAndEqual.has(propertiesList[i])) {
       i++;
     }
-    // skip all further ignored or equal characters
+
+    // Skip all further ignored or equal characters
     while (i < length && notSeparator.has(propertiesList[i])) {
       i++;
     }
-    // if we are hitting a separator, the current property has just a name and no value, move on
+
+    // If we are hitting a separator, the current property has just a name and
+    // no value, move on
     if (!sepSet.has(propertiesList[i])) {
-      // parse the value of the property
+      // Parse the value of the property
       start = i;
+
       while (i < length && !allSpecial.has(propertiesList[i])) {
         i++;
       }
+
       value = propertiesList.substring(start, i);
 
       if (value.length > 0) {
@@ -136,36 +146,51 @@ export function parsePropertiesList(
   return valueMap;
 }
 
-/*
- * Parses a "maximum-scale" property according to https://www.w3.org/TR/css-device-adapt-1/#min-scale-max-scale
+/**
+ * Parse a "maximum-scale" property.
+ *
+ * @see https://www.w3.org/TR/css-device-adapt-1/#min-scale-max-scale
  *
  * @remarks
- * This seems to be the iOS/Safari algorithm and other browsers might handle it in unknown ways.
+ * This seems to be the iOS/Safari algorithm and other browsers might handle it
+ * in unknown ways.
  */
 export function parseMaximumScale(scale: string | undefined): Option<number> {
   switch (scale) {
     case undefined:
       return None;
+
     case "yes":
       return Option.of(1);
+
     case "device-width":
     case "device-height":
       return Option.of(10);
+
     case "no":
       return Option.of(0.1);
+
     default:
       const scaleValue = Number(scale);
+
+      if (scaleValue < 0) {
+        return None;
+      }
+
       return Option.of(
         isNaN(scaleValue) ? 0.1 : Real.clamp(scaleValue, 0.1, 10)
       );
   }
 }
 
-/*
- * Parses a "user-scalable" property according to https://www.w3.org/TR/css-device-adapt-1/#user-scalable
+/**
+ * Parse a "user-scalable" property according.
+ *
+ * @see https://www.w3.org/TR/css-device-adapt-1/#user-scalable
  *
  * @remark
- * This seems to be the iOS/Safari algorithm and other browsers might handle it in unknown ways.
+ * This seems to be the iOS/Safari algorithm and other browsers might handle it
+ * in unknown ways.
  */
 export function parseUserScalable(
   scalable: string | undefined
@@ -173,14 +198,18 @@ export function parseUserScalable(
   switch (scalable) {
     case undefined:
       return None;
+
     case "yes":
     case "device-width":
     case "device-height":
       return Option.of("zoom");
+
     case "no":
       return Option.of("fixed");
+
     default:
       const scalableValue = Number(scalable);
+
       return Option.of(
         scalableValue <= -1 || scalableValue >= 1 ? "zoom" : "fixed"
       );