+++ packages/alfa-rules/src/sia-r67/rule.ts
@@ -6,11 +6,10 @@ import { Err, Ok } from "@siteimprove/alfa-result";
 import { Page } from "@siteimprove/alfa-web";
 
 import { expectation } from "../common/expectation";
-
-import { hasAccessibleName } from "../common/predicate/has-accessible-name";
+import { isMarkedDecorative } from "../common/predicate/is-marked-decorative";
 
 const { isElement, hasName, hasNamespace } = Element;
-const { and, or } = Predicate;
+const { and, or, not } = Predicate;
 
 export default Rule.Atomic.of<Page, Element>({
   uri: "https://siteimprove.github.io/sanshikan/rules/sia-r67.html",
@@ -27,7 +26,7 @@ export default Rule.Atomic.of<Page, Element>({
                   and(hasNamespace(Namespace.HTML), hasName("img")),
                   and(hasNamespace(Namespace.SVG), hasName("svg"))
                 ),
-                isMarkedAsDecorative
+                isMarkedDecorative
               )
             )
           );
@@ -36,16 +35,11 @@ export default Rule.Atomic.of<Page, Element>({
       expectations(target) {
         return {
           1: expectation(
-            Node.from(target, device).every((accNode) => {
-              const role = accNode.role();
-
-              return (
-                role.isNone() ||
-                Role.hasName("none", "presentation")(role.get())
-              );
-            }),
-            () => Outcomes.IsNotExposed,
-            () => Outcomes.IsExposed
+            Node.from(target, device).every((accNode) =>
+              accNode.role().some(not(Role.hasName("none", "presentation")))
+            ),
+            () => Outcomes.IsExposed,
+            () => Outcomes.IsNotExposed
           ),
         };
       },
@@ -55,26 +49,10 @@ export default Rule.Atomic.of<Page, Element>({
 
 export namespace Outcomes {
   export const IsNotExposed = Ok.of(
-    Diagnostic.of(
-      `The element is marked as decorative and is not exposed`
-    )
+    Diagnostic.of(`The element is marked as decorative and is not exposed`)
   );
 
   export const IsExposed = Err.of(
-    Diagnostic.of(
-      `The element is marked as decorative but is exposed`
-    )
-  );
-}
-
-/**
- * Check if an element is marked as decorative by looking at its role but without conflict resolution.
- * If the result is "none" or "presentation", then the element is marked as decorative.
- */
-function isMarkedAsDecorative(element: Element): boolean {
-  return (
-    Role.from(element, { allowPresentational: true })
-      // Element is marked as decorative if at least one browser thinks so.
-      .some((r) => r.some(Role.hasName("none", "presentation")))
+    Diagnostic.of(`The element is marked as decorative but is exposed`)
   );
 }